Refactor mock structure
