+++ frontend/packages/ux-editor/src/components/config/Expressions/SimpleExpression.test.tsx
@@ -2,7 +2,8 @@ import React from 'react';
 import { screen } from '@testing-library/react';
 import { internalExpressionWithMultipleSubExpressions } from '../../../testing/expressionMocks';
 import { ServicesContextProps } from 'app-shared/contexts/ServicesContext';
-import { formDesignerMock, renderWithMockStore } from '../../../testing/mocks';
+import { renderWithMockStore } from '../../../testing/mocks';
+import { formDesignerMock } from '../../../testing/stateMocks';
 import { SimpleExpression, SimpleExpressionProps } from './SimpleExpression';
 import { textMock } from '../../../../../../testing/mocks/i18nMock';
 import { createQueryClientMock } from 'app-shared/mocks/queryClientMock';
