+++ frontend/packages/ux-editor/src/components/config/Expressions/SubExpressionContent.test.tsx
@@ -4,7 +4,8 @@ import userEvent from '@testing-library/user-event';
 import { SubExpressionContent, SubExpressionContentProps } from './SubExpressionContent';
 import { subExpression0 } from '../../../testing/expressionMocks';
 import { ServicesContextProps } from 'app-shared/contexts/ServicesContext';
-import { formDesignerMock, renderWithMockStore } from '../../../testing/mocks';
+import { renderWithMockStore } from '../../../testing/mocks';
+import { formDesignerMock } from '../../../testing/stateMocks';
 import { textMock } from '../../../../../../testing/mocks/i18nMock';
 import { createQueryClientMock } from 'app-shared/mocks/queryClientMock';
 import { QueryKey } from 'app-shared/types/QueryKey';
