+++ frontend/packages/ux-editor/src/hooks/useText.test.ts
@@ -1,4 +1,5 @@
-import { appDataMock, renderHookWithMockStore } from '../testing/mocks';
+import { renderHookWithMockStore } from '../testing/mocks';
+import { appDataMock } from '../testing/stateMocks';
 import { useText } from './useText';
 import { IAppDataState } from '../features/appData/appDataReducers';
 import { mockUseTranslation } from '../../../../testing/mocks/i18nMock';
