+++ frontend/packages/ux-editor/src/selectors/formLayoutSelectors.test.ts
@@ -1,5 +1,5 @@
 import { IAppState, IFormLayouts } from '../types/global';
-import { appStateMock, formDesignerMock } from '../testing/mocks';
+import { appStateMock, formDesignerMock } from '../testing/stateMocks';
 import {
   getAllLayoutComponents,
   getAllLayoutContainers,
