+++ frontend/packages/ux-editor/src/testing/mocks.tsx
@@ -1,13 +1,10 @@
 import type { ReactNode } from 'react';
 import React from 'react';
 import configureStore from 'redux-mock-store';
-import type { IAppDataState } from '../features/appData/appDataReducers';
 import type { IAppState } from '../types/global';
-import type { ITextResourcesState } from '../features/appData/textResources/textResourcesSlice';
 import { Provider } from 'react-redux';
 import { render, renderHook } from '@testing-library/react';
 import { ServicesContextProps, ServicesContextProvider } from 'app-shared/contexts/ServicesContext';
-import { IFormDesignerState } from '../features/formDesigner/formDesignerReducer';
 import { ILayoutSettings } from 'app-shared/types/global';
 import { BrowserRouter } from 'react-router-dom';
 import ruleHandlerMock from './ruleHandlerMock';
@@ -19,35 +16,12 @@ import {
   layout2NameMock,
   layoutSetsMock,
 } from './layoutMock';
+import { appStateMock } from './stateMocks';
 import { queriesMock as allQueriesMock } from 'app-shared/mocks/queriesMock';
 import { QueryClient } from '@tanstack/react-query';
 import expressionSchema from './schemas/json/layout/expression.schema.v1.json';
 import numberFormatSchema from './schemas/json/layout/number-format.schema.v1.json';
 import layoutSchema from './schemas/json/layout/layout.schema.v1.json';
-
-export const textResourcesMock: ITextResourcesState = {
-  currentEditId: undefined,
-};
-
-export const appDataMock: IAppDataState = {
-  textResources: textResourcesMock,
-};
-
-export const formDesignerMock: IFormDesignerState = {
-  layout: {
-    error: null,
-    saving: false,
-    unSavedChanges: false,
-    selectedLayoutSet: 'test-layout-set',
-    selectedLayout: layout1NameMock,
-    invalidLayouts: [],
-  },
-};
-
-export const appStateMock: IAppState = {
-  appData: appDataMock,
-  formDesigner: formDesignerMock,
-};
 
 export const formLayoutSettingsMock: ILayoutSettings = {
   pages: {
@@ -59,16 +33,6 @@ export const formLayoutSettingsMock: ILayoutSettings = {
 export const textLanguagesMock = ['nb', 'nn', 'en'];
 
 export const optionListIdsMock: string[] = ['test-1', 'test-2'];
-
-export const FormContextProviderMock = {
-  formId: null,
-  form: null,
-  handleDiscard: jest.fn(),
-  handleEdit: jest.fn(),
-  handleUpdate: jest.fn(),
-  handleSave: jest.fn().mockImplementation(() => Promise.resolve()),
-  debounceSave: jest.fn().mockImplementation(() => Promise.resolve()),
-};
 
 export const queriesMock: ServicesContextProps = {
   ...allQueriesMock,
